requirements.txt
python-dotenv>=1.0.1

# GUI libraries
customtkinter>=5.2.2
darkdetect>=0.8.0
Pillow>=10.3.0

# Voice libraries
vosk>=0.3.45
sounddevice>=0.4.6
numpy>=1.26
pyttsx3>=2.90

# API libraries
wikipedia>=1.4.0
requests>=2.32.0
psutil>=5.9.0

# Testing
pytest>=8.3.0
